prover setup
