+++ crates/proof-of-sql/src/proof_primitive/dory/mod.rs
@@ -163,5 +163,3 @@ pub use dynamic_dory_commitment::DynamicDoryCommitment;
 #[cfg(test)]
 mod dynamic_dory_commitment_evaluation_proof_test;
 pub use dynamic_dory_commitment_evaluation_proof::DynamicDoryEvaluationProof;
-
-mod owned_setup;
